requirements.txt
flask==3.0.2
python-dotenv==1.0.1
werkzeug==3.0.1
python-docx==1.1.0
pyhwp==0.1b3
beautifulsoup4==4.12.3
requests==2.31.0
gunicorn==21.2.0 